Compatible con decimales

+++ app/src/main/java/com/calculadora_pacharan/MainActivity.kt
@@ -28,10 +28,10 @@ class MainActivity : AppCompatActivity() {
 
     fun calcular(){
         val campo = findViewById<EditText>(R.id.editCantAnis).text;
-        var anis = 0
+        var anis = 0.0
 
         if (!campo.toString().equals("") && campo.toString().length < 6){
-            anis = campo.toString().toInt()
+            anis = campo.toString().toDouble()
         }
 
         if (anis < 100000){
